Added !character kill command
Includes confirmation string and can only be done by the DM for security

+++ dicebot/cogs/characters.py
@@ -112,28 +112,33 @@ class CharacterCog (Cog):
             text.append(str(character))
         await ctx.send('\n'.join(text))
 
-    # @group.command()
+    @group.command()
     @commands.has_role('DM')
-    async def kill(self, ctx, *, name: str):
+    async def kill(self, ctx, name: str, *, confirmation: str):
         '''
         Deletes a character
         This is permanent and removes all associated attributes!
+        Can only be done by the @DM role
 
         Parameters:
         [name] the name of the character to delete
+        [confirmation] enter `100%` to confirm that you want to delete the character permanently
         '''
-        character = ctx.session.query(m.Character)\
-            .filter_by(name=name, server=ctx.guild.id).one_or_none()
-        if character is not None:
-            for attribute in character.attributes:
-                for item in getattr(character, attribute):
-                    ctx.session.delete(item)
-            ctx.session.commit()
-            ctx.session.delete(character)
-            ctx.session.commit()
-            await ctx.send('{} is dead'.format(str(character)))
+        if confirmation == '100%':
+            character = ctx.session.query(m.Character)\
+                .filter_by(name=name, server=ctx.guild.id).one_or_none()
+            if character is not None:
+                for attribute in character.attributes:
+                    for item in getattr(character, attribute):
+                        ctx.session.delete(item)
+                ctx.session.commit()
+                ctx.session.delete(character)
+                ctx.session.commit()
+                await ctx.send('{} is dead'.format(str(character)))
+            else:
+                await ctx.send('Error: No character named {}'.format(name))
         else:
-            await ctx.send('Error: No character named {}'.format(name))
+            await ctx.send('Error: please confirm deletion correctly')
 
     @commands.command()
     async def rest(self, ctx, *, rest: str):
